fix: follow bug
