+++ package.json
@@ -29,6 +29,7 @@
     "@koa/cors": "2.2.2",
     "axios": "0.18.0",
     "debug": "4.1.0",
+    "jsonwebtoken": "^8.3.0",
     "koa": "2.5.3",
     "koa-bodyparser": "4.2.1",
     "koa-cache-control": "2.0.0",
