+++ server/socket/constant.js
@@ -1,0 +1,14 @@
+const SYNC_TYPE = {
+  INSERT: 'create',
+  UPDATE: 'update',
+  DELETE: 'delete',
+  MOVE: 'move',
+  REDO: 'redo',
+  UNDO: 'undo',
+  MOVE_BY_PRESENTER: 'move_by_presenter',
+  ZOOM: 'zoom'
+}
+
+module.exports = {
+  SYNC_TYPE
+}
